Python_Advanced_7/my_django_project/requirements.txt
django==4.2
django-environ==0.9.0
mysqlclient==2.1.1
